% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehr_readers.R
\name{add_api_auth}
\alias{add_api_auth}
\title{Add API Authentication}
\usage{
add_api_auth(req, auth_type, auth_token, api_key, username, password)
}
\arguments{
\item{req}{httr2 request object}

\item{auth_type}{Authentication type}

\item{auth_token}{Bearer token}

\item{api_key}{API key}

\item{username}{Username for basic auth}

\item{password}{Password for basic auth}
}
\value{
Modified request object
}
\description{
Adds authentication to httr2 request
}
\keyword{internal}
