% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{AGE_RANGE}
\alias{AGE_RANGE}
\title{Standard Age Validation Range}
\format{
Numeric vector of min and max age
}
\usage{
AGE_RANGE
}
\description{
Standard Age Validation Range
}
\keyword{internal}
