% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ANALYSIS_METHODS}
\alias{ANALYSIS_METHODS}
\title{Supported Statistical Analysis Methods}
\format{
Character vector of analysis methods
}
\usage{
ANALYSIS_METHODS
}
\description{
Supported Statistical Analysis Methods
}
\keyword{internal}
