% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization.R
\name{apply_single_terminology_map}
\alias{apply_single_terminology_map}
\title{Apply Single Terminology Map}
\usage{
apply_single_terminology_map(data, map_type, map_name = NULL)
}
\arguments{
\item{data}{Data to transform}

\item{map_type}{Mapping type (e.g., "icd10_to_snomed")}
}
\value{
Transformed data
}
\description{
Applies a single terminology mapping
}
\keyword{internal}
