% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization.R
\name{apply_terminology_mappings}
\alias{apply_terminology_mappings}
\title{Apply Terminology Mappings}
\usage{
apply_terminology_mappings(data, terminology_maps)
}
\arguments{
\item{data}{Data with codes to map}

\item{terminology_maps}{Named list of mappings}
}
\value{
Data with mapped codes
}
\description{
Maps medical codes between terminology systems
}
\keyword{internal}
