% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_design.R
\name{assess_balance}
\alias{assess_balance}
\title{Assess Covariate Balance Between Two Groups}
\usage{
assess_balance(group1, group2, vars)
}
\description{
Assess Covariate Balance Between Two Groups
}
\keyword{internal}
