% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{calculate_missingness}
\alias{calculate_missingness}
\title{Calculate Missingness}
\usage{
calculate_missingness(data, vars)
}
\arguments{
\item{data}{Data frame}

\item{vars}{Variables to check}
}
\value{
Data frame with missingness statistics
}
\description{
Calculates missingness statistics for specified variables
}
\keyword{internal}
