% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_assessment.R
\name{calculate_overall_quality_score}
\alias{calculate_overall_quality_score}
\title{Calculate Overall Quality Score}
\usage{
calculate_overall_quality_score(results, dimensions)
}
\arguments{
\item{results}{List of dimension results}

\item{dimensions}{Dimensions assessed}
}
\value{
Numeric overall quality score (0-1)
}
\description{
Calculates weighted overall quality score from dimension scores
}
\keyword{internal}
