% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity_score.R
\name{calculate_ps_overlap}
\alias{calculate_ps_overlap}
\title{Calculate Propensity Score Overlap}
\usage{
calculate_ps_overlap(ps_treated, ps_control)
}
\arguments{
\item{ps_treated}{PS for treated group}

\item{ps_control}{PS for control group}
}
\value{
Overlap coefficient (0-1)
}
\description{
Calculates overlap coefficient between PS distributions
}
\keyword{internal}
