% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_error_message}
\alias{create_error_message}
\title{Create Informative Error Message}
\usage{
create_error_message(context, issue, suggestion = NULL)
}
\arguments{
\item{context}{Function or module context}

\item{issue}{Description of the issue}

\item{suggestion}{Optional suggestion for resolution}
}
\value{
Character string with formatted error message
}
\description{
Creates standardized error messages for the package
}
\keyword{internal}
