% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization.R
\name{detect_schema}
\alias{detect_schema}
\title{Detect Data Schema}
\usage{
detect_schema(data)
}
\arguments{
\item{data}{Input data frame}
}
\value{
Character schema name
}
\description{
Attempts to detect the schema/format of the input data
}
\keyword{internal}
