% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity_score.R
\name{estimate_ps_logistic}
\alias{estimate_ps_logistic}
\title{Estimate Propensity Score with Logistic Regression}
\usage{
estimate_ps_logistic(data, formula)
}
\arguments{
\item{data}{Data frame}

\item{formula}{Model formula}
}
\value{
Fitted model
}
\description{
Estimate Propensity Score with Logistic Regression
}
\keyword{internal}
