% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity_score.R
\name{estimate_ps_super_learner}
\alias{estimate_ps_super_learner}
\title{Estimate Propensity Score with Super Learner Ensemble}
\usage{
estimate_ps_super_learner(data, formula, treatment, covariates, control)
}
\description{
Estimate Propensity Score with Super Learner Ensemble
}
\keyword{internal}
