% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity_score.R
\name{estimate_ps_xgb}
\alias{estimate_ps_xgb}
\title{Estimate Propensity Score with XGBoost}
\usage{
estimate_ps_xgb(data, formula, treatment, covariates, control)
}
\description{
Estimate Propensity Score with XGBoost
}
\keyword{internal}
