% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulatory_reporting.R
\name{export_table_one_csv}
\alias{export_table_one_csv}
\title{Export Table One to CSV}
\usage{
export_table_one_csv(table_one, output_file)
}
\description{
Export Table One to CSV
}
\keyword{internal}
