% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulatory_reporting.R
\name{export_table_one_latex}
\alias{export_table_one_latex}
\title{Export Table One to LaTeX}
\usage{
export_table_one_latex(table_one, output_file, title)
}
\description{
Export Table One to LaTeX
}
\keyword{internal}
