% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehr_readers.R
\name{extract_api_data}
\alias{extract_api_data}
\title{Extract Data from API Response}
\usage{
extract_api_data(content, format)
}
\arguments{
\item{content}{Parsed JSON content}

\item{format}{Data format}
}
\value{
Data frame
}
\description{
Extracts data from JSON API response based on format
}
\keyword{internal}
