% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_data_info}
\alias{extract_data_info}
\title{Extract Data Information}
\usage{
extract_data_info(data)
}
\arguments{
\item{data}{Input data frame}
}
\value{
List with data information
}
\description{
Extracts summary information about a dataset
}
\keyword{internal}
