% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{FILE_EXTENSIONS}
\alias{FILE_EXTENSIONS}
\title{File Type Extensions}
\format{
Named list of file extensions by type
}
\usage{
FILE_EXTENSIONS
}
\description{
File Type Extensions
}
\keyword{internal}
