% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehr_readers.R
\name{filter_by_date}
\alias{filter_by_date}
\title{Filter by Date Range}
\usage{
filter_by_date(data, date_range)
}
\arguments{
\item{data}{Data frame}

\item{date_range}{Vector of start and end dates}
}
\value{
Filtered data frame
}
\description{
Filters data by date range
}
\keyword{internal}
