% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{generate_imputation_diagnostics}
\alias{generate_imputation_diagnostics}
\title{Generate Imputation Diagnostics}
\usage{
generate_imputation_diagnostics(original, imputed, vars, method)
}
\arguments{
\item{original}{Original data}

\item{imputed}{Imputed data}

\item{vars}{Variables that were imputed}

\item{method}{Imputation method}
}
\value{
List with diagnostic information
}
\description{
Generates diagnostic information about imputation
}
\keyword{internal}
