% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehr_readers.R
\name{get_main_table_for_format}
\alias{get_main_table_for_format}
\title{Get Main Table for Format}
\usage{
get_main_table_for_format(format)
}
\arguments{
\item{format}{Data format}
}
\value{
Character table name
}
\description{
Returns the main table name for a given format
}
\keyword{internal}
