% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{get_quality_threshold}
\alias{get_quality_threshold}
\title{Get Quality Threshold}
\usage{
get_quality_threshold(dimension)
}
\arguments{
\item{dimension}{Quality dimension name}
}
\value{
Numeric threshold value
}
\description{
Retrieves quality threshold for specified dimension
}
\keyword{internal}
