% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{get_schema_definition}
\alias{get_schema_definition}
\title{Get Schema Definition}
\usage{
get_schema_definition(format, version = NULL)
}
\arguments{
\item{format}{Data format name}

\item{version}{Optional version string}
}
\value{
Named list with schema definition
}
\description{
Retrieves schema definition for specified format and version
}
\keyword{internal}
