% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{IMPUTATION_METHODS}
\alias{IMPUTATION_METHODS}
\title{Supported Imputation Methods}
\format{
Character vector of imputation methods
}
\usage{
IMPUTATION_METHODS
}
\description{
Supported Imputation Methods
}
\keyword{internal}
