% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute_knn}
\alias{impute_knn}
\title{Impute with K-Nearest Neighbors}
\usage{
impute_knn(data, vars, predictors, k = 5, verbose = TRUE)
}
\arguments{
\item{data}{Data frame}

\item{vars}{Variables to impute}

\item{predictors}{Predictor variables}
}
\value{
Data frame with imputed values
}
\description{
Imputes using KNN algorithm
}
\keyword{internal}
