% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute_mean}
\alias{impute_mean}
\title{Impute Mean}
\usage{
impute_mean(data, vars)
}
\arguments{
\item{data}{Data frame}

\item{vars}{Variables to impute}
}
\value{
Data frame with imputed values
}
\description{
Imputes missing values with variable means
}
\keyword{internal}
