% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute_median}
\alias{impute_median}
\title{Impute Median}
\usage{
impute_median(data, vars)
}
\arguments{
\item{data}{Data frame}

\item{vars}{Variables to impute}
}
\value{
Data frame with imputed values
}
\description{
Imputes missing values with variable medians
}
\keyword{internal}
