% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute_mice}
\alias{impute_mice}
\title{Impute with MICE}
\usage{
impute_mice(data, vars, m, seed)
}
\arguments{
\item{data}{Data frame}

\item{vars}{Variables to impute}

\item{m}{Number of imputations}

\item{seed}{Random seed}
}
\value{
Data frame with imputed values
}
\description{
Multiple Imputation by Chained Equations
}
\keyword{internal}
