% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute_mode}
\alias{impute_mode}
\title{Impute Mode}
\usage{
impute_mode(data, vars)
}
\arguments{
\item{data}{Data frame}

\item{vars}{Variables to impute}
}
\value{
Data frame with imputed values
}
\description{
Imputes missing values with variable modes
}
\keyword{internal}
