% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute_random_forest}
\alias{impute_random_forest}
\title{Impute with Random Forest}
\usage{
impute_random_forest(data, vars, predictors)
}
\arguments{
\item{data}{Data frame}

\item{vars}{Variables to impute}

\item{predictors}{Predictor variables}
}
\value{
Data frame with imputed values
}
\description{
Imputes using random forest algorithm
}
\keyword{internal}
