% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute_xgboost}
\alias{impute_xgboost}
\title{Impute with XGBoost Models}
\usage{
impute_xgboost(
  data,
  vars,
  predictors,
  params = list(),
  nrounds = 50,
  verbose = TRUE
)
}
\description{
Impute with XGBoost Models
}
\keyword{internal}
