% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R
\name{is_rwe_quality_report}
\alias{is_rwe_quality_report}
\title{Check if Object is rwe_quality_report}
\usage{
is_rwe_quality_report(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical
}
\description{
Check if Object is rwe_quality_report
}
