% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization.R
\name{load_standard_mapping}
\alias{load_standard_mapping}
\title{Load Standard Mapping}
\usage{
load_standard_mapping(from_system, to_system)
}
\arguments{
\item{from_system}{Source system}

\item{to_system}{Target system}
}
\value{
Mapping data frame or NULL
}
\description{
Loads standard terminology mapping table
}
\keyword{internal}
