% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_error}
\alias{log_error}
\title{Log Error Message}
\usage{
log_error(message, context = NULL, ...)
}
\arguments{
\item{message}{Log message}

\item{context}{Optional context}

\item{...}{Additional objects}
}
\description{
Logs error message
}
\examples{
\dontrun{
log_error("File not found", context = "rwe_read_ehr")
}
}
