% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_info}
\alias{log_info}
\title{Log Info Message}
\usage{
log_info(message, context = NULL, ...)
}
\arguments{
\item{message}{Log message}

\item{context}{Optional context}

\item{...}{Additional objects}
}
\description{
Logs informational message
}
\examples{
\dontrun{
log_info("Data loaded successfully", context = "rwe_read_ehr")
}
}
