% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\docType{data}
\name{logging}
\alias{logging}
\alias{.logger_env}
\title{Logging Utilities}
\format{
An object of class \code{environment} of length 4.
}
\usage{
.logger_env
}
\description{
Provides logging functionality for audit trails and debugging
}
\keyword{internal}
