% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_design.R
\name{match_exact}
\alias{match_exact}
\title{Exact Matching for External Controls}
\usage{
match_exact(external_df, treatment_df, matching_vars, ratio)
}
\description{
Exact Matching for External Controls
}
\keyword{internal}
