% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_design.R
\name{match_nearest}
\alias{match_nearest}
\title{Nearest Neighbor Matching for External Controls}
\usage{
match_nearest(external_df, treatment_df, matching_vars, ratio, replace)
}
\description{
Nearest Neighbor Matching for External Controls
}
\keyword{internal}
