% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_design.R
\name{match_propensity_score}
\alias{match_propensity_score}
\title{Propensity Score Matching for External Controls}
\usage{
match_propensity_score(
  external_df,
  treatment_df,
  matching_vars,
  ratio,
  caliper,
  replace
)
}
\description{
Propensity Score Matching for External Controls
}
\keyword{internal}
