% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{MATCHING_METHODS}
\alias{MATCHING_METHODS}
\title{Supported Matching Methods}
\format{
Character vector of matching methods
}
\usage{
MATCHING_METHODS
}
\description{
Supported Matching Methods
}
\keyword{internal}
