% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R
\name{new_rwe_analysis}
\alias{new_rwe_analysis}
\title{Create rwe_analysis Object}
\usage{
new_rwe_analysis(primary_analysis, method, class_name = NULL)
}
\arguments{
\item{primary_analysis}{Primary analysis results}

\item{method}{Analysis method used}

\item{class_name}{Additional class name(s)}
}
\value{
Object of class rwe_analysis
}
\description{
Base class for analysis results
}
