% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R
\name{new_rwe_data}
\alias{new_rwe_data}
\title{Create rwe_data Object}
\usage{
new_rwe_data(data, metadata = list(), class_name = NULL)
}
\arguments{
\item{data}{Data frame or tibble}

\item{metadata}{List of metadata}

\item{class_name}{Additional class name(s)}
}
\value{
Object of class rwe_data
}
\description{
Base class for all data objects in the package
}
\keyword{internal}
