% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R
\name{new_rwe_ehr}
\alias{new_rwe_ehr}
\title{Create rwe_ehr Object}
\usage{
new_rwe_ehr(data, metadata = list(), schema = list())
}
\arguments{
\item{data}{Data frame containing EHR data}

\item{metadata}{List of metadata}

\item{schema}{Schema definition}
}
\value{
Object of class rwe_ehr
}
\description{
Class for EHR data with schema information
}
\examples{
\dontrun{
ehr_data <- new_rwe_ehr(
  data = patient_data,
  metadata = list(source = "hospital_ehr", format = "omop_cdm"),
  schema = omop_schema
)
}
}
