% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R
\name{new_rwe_harmonized}
\alias{new_rwe_harmonized}
\title{Create rwe_harmonized Object}
\usage{
new_rwe_harmonized(
  data,
  source_schema,
  target_schema,
  mapping_config,
  validation = NULL
)
}
\arguments{
\item{data}{Harmonized data frame}

\item{source_schema}{Source schema name}

\item{target_schema}{Target schema name}

\item{mapping_config}{Mapping configuration used}

\item{validation}{Validation results}
}
\value{
Object of class rwe_harmonized
}
\description{
Class for harmonized/transformed data
}
