% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R
\name{new_rwe_matched_cohort}
\alias{new_rwe_matched_cohort}
\title{Create rwe_matched_cohort Object}
\usage{
new_rwe_matched_cohort(
  matched_data,
  match_object,
  balance_before,
  balance_after
)
}
\arguments{
\item{matched_data}{Matched data frame}

\item{match_object}{Matching object from MatchIt}

\item{balance_before}{Balance before matching}

\item{balance_after}{Balance after matching}
}
\value{
Object of class rwe_matched_cohort
}
\description{
Class for propensity score matched cohorts
}
