% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{OMOP_CDM_5.4_SCHEMA}
\alias{OMOP_CDM_5.4_SCHEMA}
\title{OMOP CDM 5.4 Core Tables}
\format{
Named list of OMOP CDM table schemas
}
\usage{
OMOP_CDM_5.4_SCHEMA
}
\description{
OMOP CDM 5.4 Core Tables
}
\keyword{internal}
