% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{PACKAGE_METADATA}
\alias{PACKAGE_METADATA}
\title{Package Version Information}
\format{
List with version metadata
}
\usage{
PACKAGE_METADATA
}
\description{
Package Version Information
}
\keyword{internal}
