% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_analysis.R
\name{plot.rwe_km}
\alias{plot.rwe_km}
\title{Plot Method for Kaplan-Meier (Survival Curves)}
\usage{
\method{plot}{rwe_km}(x, ...)
}
\arguments{
\item{x}{rwe_km object}

\item{...}{Additional arguments}
}
\description{
Plot Method for Kaplan-Meier (Survival Curves)
}
