% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_analysis.R
\name{plot.rwe_survival}
\alias{plot.rwe_survival}
\title{Plot Method for Cox Regression (Forest Plot)}
\usage{
\method{plot}{rwe_survival}(x, ...)
}
\arguments{
\item{x}{rwe_survival object}

\item{...}{Additional arguments}
}
\description{
Plot Method for Cox Regression (Forest Plot)
}
