% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity_score.R
\name{predict.ps_rf_model}
\alias{predict.ps_rf_model}
\title{Predict Method for RF Propensity Score Model}
\usage{
\method{predict}{ps_rf_model}(object, newdata = NULL, type = "response", ...)
}
\arguments{
\item{object}{ps_rf_model object}

\item{newdata}{New data for prediction}

\item{type}{Type of prediction}

\item{...}{Additional arguments}
}
\value{
Predicted probabilities
}
\description{
Predict Method for RF Propensity Score Model
}
\keyword{internal}
