% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_inference.R
\name{print.rwe_ate}
\alias{print.rwe_ate}
\title{Print Method for ATE Object}
\usage{
\method{print}{rwe_ate}(x, ...)
}
\arguments{
\item{x}{rwe_ate object}

\item{...}{Additional arguments}
}
\description{
Print Method for ATE Object
}
