% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{print.rwe_audit_trail}
\alias{print.rwe_audit_trail}
\title{Print Audit Trail}
\usage{
\method{print}{rwe_audit_trail}(x, ...)
}
\arguments{
\item{x}{Audit trail object}

\item{...}{Additional arguments}
}
\description{
Print Audit Trail
}
