% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulatory_reporting.R
\name{print.rwe_consort}
\alias{print.rwe_consort}
\title{Print Method for CONSORT Diagrams}
\usage{
\method{print}{rwe_consort}(x, ...)
}
\arguments{
\item{x}{An object of class \code{rwe_consort}}

\item{...}{Additional arguments (not used)}
}
\description{
Print Method for CONSORT Diagrams
}
