% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_inference.R
\name{print.rwe_doubly_robust}
\alias{print.rwe_doubly_robust}
\title{Print Method for Doubly Robust Object}
\usage{
\method{print}{rwe_doubly_robust}(x, ...)
}
\arguments{
\item{x}{rwe_doubly_robust object}

\item{...}{Additional arguments}
}
\description{
Print Method for Doubly Robust Object
}
