% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_design.R
\name{print.rwe_external_control}
\alias{print.rwe_external_control}
\title{Print Method for External Control Objects}
\usage{
\method{print}{rwe_external_control}(x, ...)
}
\arguments{
\item{x}{An object of class \code{rwe_external_control}}

\item{...}{Additional arguments (not used)}
}
\description{
Print Method for External Control Objects
}
