% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{print.rwe_imputation}
\alias{print.rwe_imputation}
\title{Print Method for Imputation Object}
\usage{
\method{print}{rwe_imputation}(x, ...)
}
\arguments{
\item{x}{rwe_imputation object}

\item{...}{Additional arguments}
}
\description{
Print Method for Imputation Object
}
