% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_inference.R
\name{print.rwe_iptw}
\alias{print.rwe_iptw}
\title{Print Method for IPTW Object}
\usage{
\method{print}{rwe_iptw}(x, ...)
}
\arguments{
\item{x}{rwe_iptw object}

\item{...}{Additional arguments}
}
\description{
Print Method for IPTW Object
}
