% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_analysis.R
\name{print.rwe_km}
\alias{print.rwe_km}
\title{Print Method for Kaplan-Meier Results}
\usage{
\method{print}{rwe_km}(x, ...)
}
\arguments{
\item{x}{rwe_km object}

\item{...}{Additional arguments}
}
\description{
Print Method for Kaplan-Meier Results
}
