% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity_score.R
\name{print.rwe_propensity_score}
\alias{print.rwe_propensity_score}
\title{Print Method for Propensity Score Object}
\usage{
\method{print}{rwe_propensity_score}(x, ...)
}
\arguments{
\item{x}{rwe_propensity_score object}

\item{...}{Additional arguments}
}
\description{
Print Method for Propensity Score Object
}
