% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulatory_reporting.R
\name{print.rwe_regulatory_report}
\alias{print.rwe_regulatory_report}
\title{Print Method for Regulatory Reports}
\usage{
\method{print}{rwe_regulatory_report}(x, ...)
}
\arguments{
\item{x}{An object of class \code{rwe_regulatory_report}}

\item{...}{Additional arguments (not used)}
}
\description{
Print Method for Regulatory Reports
}
