% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulatory_reporting.R
\name{print.rwe_safety_report}
\alias{print.rwe_safety_report}
\title{Print Method for Safety Reports}
\usage{
\method{print}{rwe_safety_report}(x, ...)
}
\arguments{
\item{x}{An object of class \code{rwe_safety_report}}

\item{...}{Additional arguments (not used)}
}
\description{
Print Method for Safety Reports
}
