% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_design.R
\name{print.rwe_study_design}
\alias{print.rwe_study_design}
\title{Print Method for Study Design Objects}
\usage{
\method{print}{rwe_study_design}(x, ...)
}
\arguments{
\item{x}{An object of class \code{rwe_study_design}}

\item{...}{Additional arguments (not used)}
}
\description{
Print Method for Study Design Objects
}
