% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_analysis.R
\name{print.rwe_survival}
\alias{print.rwe_survival}
\title{Print Method for Cox Regression Results}
\usage{
\method{print}{rwe_survival}(x, ...)
}
\arguments{
\item{x}{rwe_survival object}

\item{...}{Additional arguments}
}
\description{
Print Method for Cox Regression Results
}
