% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulatory_reporting.R
\name{print.rwe_table_one}
\alias{print.rwe_table_one}
\title{Print Method for Table One}
\usage{
\method{print}{rwe_table_one}(x, ...)
}
\arguments{
\item{x}{An object of class \code{rwe_table_one}}

\item{...}{Additional arguments (not used)}
}
\description{
Print Method for Table One
}
