% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization.R
\name{print.rwe_terminology_mapping}
\alias{print.rwe_terminology_mapping}
\title{Print Method for Terminology Mapping}
\usage{
\method{print}{rwe_terminology_mapping}(x, ...)
}
\arguments{
\item{x}{rwe_terminology_mapping object}

\item{...}{Additional arguments}
}
\description{
Print Method for Terminology Mapping
}
