% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{PS_METHODS}
\alias{PS_METHODS}
\title{Supported Propensity Score Methods}
\format{
Character vector of propensity score estimation methods
}
\usage{
PS_METHODS
}
\description{
Supported Propensity Score Methods
}
\keyword{internal}
