% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{QUALITY_COLORS}
\alias{QUALITY_COLORS}
\title{Color Palette for Quality Reports}
\format{
Named character vector of colors
}
\usage{
QUALITY_COLORS
}
\description{
Color Palette for Quality Reports
}
\keyword{internal}
