% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{quality_thresholds}
\alias{quality_thresholds}
\alias{DEFAULT_QUALITY_THRESHOLDS}
\title{Package Constants and Configurations}
\format{
Named list of quality dimension thresholds
}
\usage{
DEFAULT_QUALITY_THRESHOLDS
}
\description{
Defines standard constants, schemas, and configurations used throughout the package
}
\keyword{Data}
\keyword{Default}
\keyword{Quality}
\keyword{Thresholds}
\keyword{internal}
