% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehr_readers.R
\name{read_csv_file}
\alias{read_csv_file}
\title{Read CSV File}
\usage{
read_csv_file(file, ...)
}
\arguments{
\item{file}{File path}

\item{...}{Additional arguments passed to read function}
}
\value{
Data frame
}
\description{
Reads CSV file using readr or data.table
}
\keyword{internal}
