% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehr_readers.R
\name{read_ehr_api}
\alias{read_ehr_api}
\title{Read EHR API}
\usage{
read_ehr_api(url, format, ...)
}
\arguments{
\item{url}{API endpoint URL}

\item{format}{Data format}

\item{...}{Additional arguments}
}
\value{
Data frame
}
\description{
Reads EHR data from API endpoint (placeholder for future implementation)
}
\keyword{internal}
