% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehr_readers.R
\name{read_ehr_database}
\alias{read_ehr_database}
\title{Read EHR Database}
\usage{
read_ehr_database(connection, format, schema = NULL, ...)
}
\arguments{
\item{connection}{DBI connection object}

\item{format}{Data format}

\item{schema}{Database schema name}

\item{...}{Additional arguments}
}
\value{
Data frame or dbplyr lazy tibble
}
\description{
Reads EHR data from database connection
}
\keyword{internal}
