% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehr_readers.R
\name{read_ehr_file}
\alias{read_ehr_file}
\title{Read EHR File}
\usage{
read_ehr_file(source, format, lazy = FALSE, ...)
}
\arguments{
\item{source}{File path}

\item{format}{Data format}

\item{lazy}{Use lazy evaluation}

\item{...}{Additional arguments}
}
\value{
Data frame or arrow dataset
}
\description{
Reads EHR data from file (CSV, Parquet, Arrow, etc.)
}
\keyword{internal}
