% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehr_readers.R
\name{read_feather_file}
\alias{read_feather_file}
\title{Read Feather File}
\usage{
read_feather_file(file, ...)
}
\arguments{
\item{file}{File path}

\item{...}{Additional arguments}
}
\value{
Data frame
}
\description{
Reads Feather file using arrow package
}
\keyword{internal}
