% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehr_readers.R
\name{read_parquet_file}
\alias{read_parquet_file}
\title{Read Parquet File}
\usage{
read_parquet_file(file, lazy = FALSE, ...)
}
\arguments{
\item{file}{File path}

\item{lazy}{Use lazy evaluation}

\item{...}{Additional arguments}
}
\value{
Data frame or arrow dataset
}
\description{
Reads Parquet file using arrow package
}
\keyword{internal}
