% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{REGULATORY_TEMPLATES}
\alias{REGULATORY_TEMPLATES}
\title{Available Regulatory Report Templates}
\format{
Character vector of report templates
}
\usage{
REGULATORY_TEMPLATES
}
\description{
Available Regulatory Report Templates
}
\keyword{internal}
