% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{REPORT_FORMATS}
\alias{REPORT_FORMATS}
\title{Report Output Formats}
\format{
Character vector of output formats
}
\usage{
REPORT_FORMATS
}
\description{
Report Output Formats
}
\keyword{internal}
