% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_analysis.R
\name{rwe_kaplan_meier}
\alias{rwe_kaplan_meier}
\title{Kaplan-Meier Survival Curves}
\usage{
rwe_kaplan_meier(data, time_var, event_var, group_var = NULL, weights = NULL)
}
\arguments{
\item{data}{Study cohort data}

\item{time_var}{Name of time-to-event variable}

\item{event_var}{Name of event indicator variable}

\item{group_var}{Grouping variable (e.g., treatment) for stratification}

\item{weights}{Optional IPTW weights}
}
\value{
rwe_km object with survival curves and comparison tests
}
\description{
Estimates Kaplan-Meier survival curves, optionally stratified by treatment.
Performs log-rank test to compare survival between groups.
}
\examples{
\dontrun{
km_result <- rwe_kaplan_meier(
  data = study_data,
  time_var = "survival_days",
  event_var = "died",
  group_var = "treatment"
)

plot(km_result)
}

}
