% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{STANDARD_TERMINOLOGY_MAPS}
\alias{STANDARD_TERMINOLOGY_MAPS}
\title{Standard Terminology Mapping Pairs}
\format{
Named list of common terminology mappings
}
\usage{
STANDARD_TERMINOLOGY_MAPS
}
\description{
Standard Terminology Mapping Pairs
}
\keyword{internal}
