% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{STANDARD_TERMINOLOGY_TABLES}
\alias{STANDARD_TERMINOLOGY_TABLES}
\title{Standard Terminology Mapping Tables}
\format{
Named list of small reference mapping tables for validation/testing
}
\usage{
STANDARD_TERMINOLOGY_TABLES
}
\description{
Standard Terminology Mapping Tables
}
\keyword{internal}
