% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization.R
\name{standardize_datetime}
\alias{standardize_datetime}
\title{Standardize DateTime}
\usage{
standardize_datetime(data, target_def)
}
\arguments{
\item{data}{Input data}

\item{target_def}{Target schema definition}
}
\value{
Data with standardized dates
}
\description{
Standardizes date and datetime columns to consistent format
}
\keyword{internal}
