% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulatory_reporting.R
\name{summary.rwe_consort}
\alias{summary.rwe_consort}
\title{Summary Method for CONSORT Diagrams}
\usage{
\method{summary}{rwe_consort}(object, ...)
}
\arguments{
\item{object}{An object of class \code{rwe_consort}}

\item{...}{Additional arguments (not used)}
}
\description{
Summary Method for CONSORT Diagrams
}
