% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_design.R
\name{summary.rwe_external_control}
\alias{summary.rwe_external_control}
\title{Summary Method for External Control Objects}
\usage{
\method{summary}{rwe_external_control}(object, ...)
}
\arguments{
\item{object}{An object of class \code{rwe_external_control}}

\item{...}{Additional arguments (not used)}
}
\description{
Summary Method for External Control Objects
}
