% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulatory_reporting.R
\name{summary.rwe_safety_report}
\alias{summary.rwe_safety_report}
\title{Summary Method for Safety Reports}
\usage{
\method{summary}{rwe_safety_report}(object, ...)
}
\arguments{
\item{object}{An object of class \code{rwe_safety_report}}

\item{...}{Additional arguments (not used)}
}
\description{
Summary Method for Safety Reports
}
