% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_design.R
\name{summary.rwe_study_design}
\alias{summary.rwe_study_design}
\title{Summary Method for Study Design Objects}
\usage{
\method{summary}{rwe_study_design}(object, ...)
}
\arguments{
\item{object}{An object of class \code{rwe_study_design}}

\item{...}{Additional arguments (not used)}
}
\description{
Summary Method for Study Design Objects
}
