% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulatory_reporting.R
\name{summary.rwe_table_one}
\alias{summary.rwe_table_one}
\title{Summary Method for Table One}
\usage{
\method{summary}{rwe_table_one}(object, ...)
}
\arguments{
\item{object}{An object of class \code{rwe_table_one}}

\item{...}{Additional arguments (not used)}
}
\description{
Summary Method for Table One
}
