% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SUPPORTED_DATA_FORMATS}
\alias{SUPPORTED_DATA_FORMATS}
\title{Supported Data Source Formats}
\format{
Character vector of supported data formats
}
\usage{
SUPPORTED_DATA_FORMATS
}
\description{
Supported Data Source Formats
}
\keyword{internal}
