% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SUPPORTED_TERMINOLOGIES}
\alias{SUPPORTED_TERMINOLOGIES}
\title{Supported Medical Terminology Systems}
\format{
Character vector of supported terminology systems
}
\usage{
SUPPORTED_TERMINOLOGIES
}
\description{
Supported Medical Terminology Systems
}
\keyword{internal}
