% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{TREATMENT_COLORS}
\alias{TREATMENT_COLORS}
\title{Treatment Group Colors}
\format{
Character vector of colors for treatment groups
}
\usage{
TREATMENT_COLORS
}
\description{
Treatment Group Colors
}
\keyword{internal}
