% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{UNIT_CONVERSIONS}
\alias{UNIT_CONVERSIONS}
\title{Measurement Unit Conversion Table}
\format{
Data frame of common unit conversions with multiplicative factors
}
\usage{
UNIT_CONVERSIONS
}
\description{
Measurement Unit Conversion Table
}
\keyword{internal}
