% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{VALID_DATE_RANGE}
\alias{VALID_DATE_RANGE}
\title{Date Range for Valid Healthcare Data}
\format{
Date vector of min and max dates
}
\usage{
VALID_DATE_RANGE
}
\description{
Date Range for Valid Healthcare Data
}
\keyword{internal}
