% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehr_readers.R
\name{validate_ehr_schema}
\alias{validate_ehr_schema}
\title{Validate EHR Schema}
\usage{
validate_ehr_schema(data, format, version)
}
\arguments{
\item{data}{Data frame to validate}

\item{format}{Data format}

\item{version}{Format version}
}
\value{
List with validation results
}
\description{
Validates data against expected EHR schema
}
\keyword{internal}
