% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{write_log_header}
\alias{write_log_header}
\title{Write Log Header}
\usage{
write_log_header(file, append)
}
\arguments{
\item{file}{Log file path}

\item{append}{Whether to append}
}
\description{
Writes header information to log file
}
\keyword{internal}
